requirements.txt
streamlit
requests
pymysql
sqlalchemy
pandas
mysql-connector-python